requirements.txt
fastapi==0.109.2
Flask==3.0.2
matplotlib==3.8.2
networkx==3.2.1
Pillow==10.1.0
Pillow==10.2.0
protobuf==4.25.3
python-dotenv==1.0.1
streamlit==1.31.1
